Move makeFieldValueGetter from Policies to entityStore.ts.
It was always strange to put this logic in the Policies class. Now that
the EntityStore knows about Policies, the balance of responsibility has
naturally shifted to EntityStore.
